LayoutTests/platform/mac/accessibility/selected-text-range-for-empty-textarea.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<textarea id="textarea" tabindex=0></textarea>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the AXSelectedTextRange for a text area will be 0, 0 instead of nil.");

    if (window.accessibilityController) {

          // this text field should be required.
          document.getElementById("textarea").focus();
          var obj = accessibilityController.focusedElement;

          shouldBe("obj.selectedTextRange", "'{0, 0}'");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
